pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuroloopy"
version = "0.1.0"
description = "Real-time fMRI neurofeedback package"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["fmri", "neurofeedback", "real-time", "neuroscience"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "watchdog>=2.1.0",
    "nibabel>=3.0.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "requests>=2.25.0",
    "pyyaml>=5.4.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
]

[project.scripts]
neuroloopy = "neuroloopy.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/neuroloopy"
Repository = "https://github.com/yourusername/neuroloopy"
Documentation = "https://github.com/yourusername/neuroloopy#readme"
